adderall.dsl: Turn fresh into a macro, and spice it up
This changes the syntax of fresh: instead of returning an LVar or a list
of LVars, it can now also run goals. If called without goals, works as
before (but takes a list of symbols instead of a space separated
string). If called with goals, wraps them in allᵍ, while also
introducing the fresh vars.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>