adderall.lvar: Introduce (unbound) and better unbound marking in LVars
If an LVar is unbound, mark it as such with an internal flag, so we
don't have to rely on the name to figure that out. While there, also
introduce an (unbound) function to adderall.lvar, and remove it from the
tests and a similar one from adderall.internal.

This cleans up the code a little.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>