adderall.internal: Remove the bogus cons-substitute hack
This hack was just an attempt to fix a bug, which it did, but caused
others, and hid the real bug. Now that the real bug has been fixed along
the way, this bogus hack can be removed.

This enabled us to turn on a couple of previously disabled tests.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>