tests: Fixed and enabled another test
There were two typos in the test for frame 4.57: in the results, an #U2
was mistakenly written as #U3, and in the test itself, w was
accidentally written as :w. The second problem resulted in the query
returning only unbound values, while the first obviously made the test
fail after the other issue was fixed.

This patch corrects both mistakes.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>