adderall.bitnum: Restore python3 support
Because Hy does not do punycoding of UTF stuffs on python3, -ᵒ ends up
being changed to _ᵒ, and functions that start with an underscore do not
get imported when doing a *-import. To work around this, explicitly list
all the symbols we want to export, but only on python3. Python2 handles
it well, and function names are different there.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>