adderall: Drop the adderall.lvar module
Most of the things that were in the adderall.lvar module have been moved
to the adderall.internal module instead, and the remaining few things to
adderall.dsl. Reason being, these things are either private, and thus
belong to adderall.internal, or public, in which case their place is in
adderall.dsl.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>