Replace `cons`, `cons?`, `cdr` and work around `islice` wrapper error.
Regrettably, all iterators returned by `rest` are explicitly evaluated as lists.
This was done to work around Hy issues involving `islice` in macros, and it needs
to re-assessed.

Also, now that there are no built-in cons cells, all the logic that used them
needs to be reworked.  This commit only makes minor changes that allow the code
to compile.